write reform

+++ src/utils/common_utils.c
@@ -24,11 +24,11 @@ void exit_with_error(char* error_format, ...){
 }
 
 // https://stackoverflow.com/questions/47745149/reverse-an-integer-array-in-c
-void ReverseArray(int arr[], int size)
+void ReverseArray(uint8_t arr[], int size)
 {
     for (int i = 0; i < size/2; ++i)
     {
-        int temp = arr[i];
+        uint8_t temp = arr[i];
         arr[i] = arr[size - 1 - i];
         arr[size - 1 - i] = temp;
     }
